authlogin: Allow reading MOTD files and symlinks
Add a new type `pam_motd_var_run_t` which allows authlogin programs
such as sshd to read files and symlinks to files of this type.

`/run/motd` and `/run/motd.d/{*}` are labelled as the new type,
which sshd reads from by default to display MOTDs. Services and users
are allowed to edit files in these locations, same as any other file,
provided they have the necessary permissions. This is the same
behaviour to the outside as when the directories were labelled
`pam_var_run_t`.